crawler/requirements.txt
elasticsearch
selenium
pillow
bs4
argparse
requests
